refactor: Replace hardcoded task names with variables
- Add s_task_name variable in all task files for watchdog registration
- Replace hardcoded strings in rx_iwdt_register_task() calls
- Replace hardcoded strings in rx_iwdt_task_heartbeat() calls
- Fix unsigned long cast in rx_iwdt.c to use PRIu32 format specifier

Files modified:
- lib/rx_hal/src/rx_iwdt.c: Use PRIu32 instead of unsigned long
- src/tasks/comm_manager.c: Add s_task_name = "Comm_Manager"
- src/tasks/env_monitor.c: Add s_task_name = "Env_Monitor"
- src/tasks/motor_controller.c: Add s_task_name = "Motor_Controller"
- src/tasks/system_health.c: Add s_task_name = "System_Health"

+++ star-rx72n-firmware/lib/rx_hal/src/rx_iwdt.c
@@ -301,6 +301,7 @@ void rx_iwdt_clear_status(void)
  * =============================================================================
  */
 
+#include <inttypes.h>
 #include <stdio.h>
 #include <string.h>
 
@@ -413,9 +414,9 @@ rx_err_t rx_iwdt_check_tasks(void)
       char           log_msg[k_log_msg_buffer_size];
       const uint32_t written = (uint32_t)snprintf(log_msg,
                                                    sizeof(log_msg),
-                                                   "Task deadlock: %s (timeout %lu ms)",
+                                                   "Task deadlock: %s (timeout %" PRIu32 " ms)",
                                                    task->task_name,
-                                                   (unsigned long)task->timeout_ms);
+                                                   task->timeout_ms);
 
       if (written > 0 && written < sizeof(log_msg)) {
         rx_log_error(s_tag, log_msg);
